Finish refactoring server/services
