+++ src/app/api/validate-token/route.ts
@@ -1,4 +1,4 @@
-import { validateAccessToken } from '@/server/services/validateTokenService'
+import { validateAccessToken } from '@/server/services/validateToken'
 import { AUTH_NO_TOKEN_ERROR_RES } from '@/server/utils/authHeaders'
 import { SHARED_APIFields__ValidateToken } from '@/shared/models'
