+++ src/server/services/validateToken/index.ts
